refactor: pytagi utils
